+++ CHANGELOG.md
@@ -3,7 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.9.1] - 2023-10-19 WangXH <kurapica125@outlook.com>
 ### Changed
-- The reactive system is re-designed, `System.ISubscription` is added to provide features like `OnUnsubscribe` event, `IsUnsubscribed` property and `Unsubscribe` method, all removed form the `System.IObserver`.
+- The reactive system is re-designed, `System.ISubscription` is added for disposing the subscription, it provide features like `OnUnsubscribe` event, `IsUnsubscribed` property, those features all removed form the `System.IObserver`.
 
 
 ## [1.6.37] - 2021-08-09 WangXH <kurapica125@outlook.com>